% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{capitalize_string}
\alias{capitalize_string}
\title{Capitalize string}
\usage{
capitalize_string(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
The first element of the input vector capitalized.
}
\description{
This function takes a character vector as input and returns the capitalized version of the first element.
}
