% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find_average}
\alias{find_average}
\title{Find average}
\usage{
find_average(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The average of the input vector.
}
\description{
This function takes a numeric vector as input and returns its average.
}
