% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multiply_two_numbers}
\alias{multiply_two_numbers}
\title{Multiply two numbers}
\usage{
multiply_two_numbers(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The product of x and y.
}
\description{
This function takes two numbers as input and returns their product.
}
