% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reverse_vector}
\alias{reverse_vector}
\title{Reverse vector}
\usage{
reverse_vector(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The reversed version of the input vector.
}
\description{
This function takes a vector as input and returns its reverse.
}
